% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{bw2fft}
\alias{bw2fft}
\title{bw2fft}
\usage{
bw2fft(bwpath, fftpath = gsub("(\\\\.bw.*)|(\\\\.bigwig.*)", ".rds", bwpath),
  region = NULL, chrsub = TRUE, verbose = FALSE, buffer = 1e+05,
  skip.sweep = FALSE, vmode = "double", resume = FALSE,
  min.gapwidth = 1000)
}
\arguments{
\item{bwpath}{path to BigWig}

\item{fftpath}{path to ffTrack .rds that will be created by this (by default .bw is replaced by .rds)}

\item{region}{GRanges specifying regions to limit ffTrack computation to (default is whole genome, ie seqnames of BigWig file)}

\item{chrsub}{whether to sub out 'chr' in seqnames / seqlevels of BigWig object}

\item{verbose}{logical flag}

\item{buffer}{integer size of how big of a buffer to use when transferring data from BigWig to ffTrack object}

\item{skip.sweep}{logical flag (default FALSE) if TRUE will skip the sweep of "region" for the portions that have non-NA values}

\item{vmode}{character specifyhing vmode to use for encoding (by default double)}

\item{resume}{logical flag specifying whether to resume the populatino of an already existing ffTrack object (default FALSE)}

\item{min.gapwidth}{minimum gap-width with which to merge reference adjacent intervals, this will mildly increase the file size but reduce the range complexity of the GRanges object}
}
\value{
ffTrack object corresponding to the data in the BigWig file
}
\description{
Creates ffTrack object from input bigwig file.
}
\author{
Marcin Imielinski
}
