% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{cp}
\alias{cp}
\title{cp

copy ffTrack object to new path on the file system (and all data files)}
\usage{
cp(.Object, path, overwrite = F)
}
\description{
cp

copy ffTrack object to new path on the file system (and all data files)
}
\author{
Marcin Imielinski
}

