% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{del}
\alias{del}
\title{del

delete ffTrack object (and all data files)}
\usage{
del(.Object, path, overwrite = F)
}
\description{
del

delete ffTrack object (and all data files)
}
\author{
Marcin Imielinski
}

