% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\docType{class}
\name{ffTrack-class}
\alias{ffTrack-class}
\title{ffTrack-class}
\description{
class::ffTrack

Class \code{ffTrack} a pointer for rapid GRanges-based access to genomic data on disk.
}
\author{
Marcin Imielinski
}

