% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack-package.R
\docType{package}
\name{ffTrack-package}
\alias{ffTrack-package}
\title{ffTrack: GenomicRanges-based R package for fast, out-of-core memory access and storage of genomic data using ff}
\description{
ffTrack allows building of objects of large genomically-indexed base- resolution sequence or numeric data for quasi-random out of memory access. ffTrack objects are indexable using GRanges from GenomicRanges package. Underlying files are written as .ffData objects using the ff package.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/mskilab/ffTrack/issues}
}

}
