% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{mv}
\alias{mv}
\title{mv

moves location of .Object to new filepath, returns new updated object}
\usage{
mv(.Object, path, overwrite = F, keep.original = F)
}
\description{
mv

moves location of .Object to new filepath, returns new updated object
}
\author{
Marcin Imielinski
}

