% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\docType{methods}
\name{seqlevels}
\alias{seqlevels}
\title{description

seqlevels of ffTrack object}
\usage{
\S4method{seqlevels}{ffTrack}(x)
}
\description{
description

seqlevels of ffTrack object
}
\author{
Marcin Imielinski
}

