% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{[}
\alias{[}
\title{[}
\usage{
\S4method{[}{ffTrack,ANY,ANY,ANY}(x, i, j, gr = FALSE, raw = FALSE, drop = FALSE)
}
\arguments{
\item{i}{GRanges or GRangesList}

\item{gr}{logical flag whether to return GRanges with field $Score populated with values (=FALSE)}

\item{raw}{logical flag whether to convert raw data to levels (if levels exist) (=FALSE)}
}
\value{
either vector (if i is a GRanges) or vector list (if i is a GRangesList)
}
\description{
Takes as input either GRanges or GRangesList "i", and returns a vector or vector list (respectively) of
data from the corresponding ranges.

Strand is taken into account here - i.e. a negative range will yield reversed data (note: not reverse complement)
}
\author{
Marcin Imielinski
}
