% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{writeable<-}
\alias{writeable<-}
\title{toggle writeable status of ffTrack object}
\usage{
writeable(.Object) <- value
}
\description{
Toggle writeable status of ffTrack object
}
\author{
Marcin Imielinski
}
