% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{writeable}
\alias{writeable}
\title{writeable

access writeable status of ffTrack object}
\usage{
writeable(.Object)
}
\description{
writeable

access writeable status of ffTrack object
}
\author{
Marcin Imielinski
}

