% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{writeable_status}
\alias{writeable_status}
\title{toggle writeable status of ffTrack object}
\usage{
writeable_status(.Object, value)
}
\description{
Toggle writeable status of ffTrack object
}
\author{
Marcin Imielinski
}
