% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{cp}
\alias{cp}
\title{copy ffTrack object to new path on the file system (and all data files)}
\usage{
cp(.Object, path, overwrite = FALSE)
}
\description{
copy ffTrack object to new path on the file system (and all data files)
}
\author{
Marcin Imielinski
}
