% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{get_seq}
\alias{get_seq}
\title{get_seq}
\usage{
get_seq(hg, gr, unlist = TRUE, mc.cores = 1, mc.chunks = mc.cores,
  add.chr = FALSE, as.data.table = FALSE, verbose = FALSE)
}
\arguments{
\item{hg}{A BSgenome or and ffTrack object with levels = c('A','T','G','C','N')}

\item{gr}{GRanges object to define the ranges}

\item{unlist}{logical whether to unlist the final output into a single DNAStringSet. (default = TRUE)}

\item{mc.cores}{Optional multicore call. (default = 1)}

\item{mc.chunks}{Optional define how to chunk the multicore call. (default = mc.cores)}

\item{as.data.table}{boolean (default = FALSE)}

\item{verbose}{boolean Flag to increase verbosity (default = FALSE)}
}
\value{
DNAStringSet of sequences
}
\description{
Retrieve genomic sequences

Wrapper around getSeq which does the "chr" and seqnames conversion if necessary
also handles GRangesList queries
}
