% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffTrack.R
\name{mv}
\alias{mv}
\title{mv}
\usage{
mv(.Object, path, overwrite = FALSE, keep.original = FALSE)
}
\description{
moves location of .Object to new filepath, returns new updated object
}
\author{
Marcin Imielinski
}

